[OU-FIX] stock: fill picking type default locations in end-migration
Other modules like stock_dropshipping and repair customize the default.
